refact: extrair toggle expand/collapse para ItemExpand (actor-sheet.mjs)
